PR review pass 9
- Renamed PowerSystemNode to PowerModelHandler.

- Renamed PowerSystemPack to PowerSystemNode, to better represent this being the ROS node for the power system.

- Made a pass through all files to rename any relevant variables/parameters/comments/etc to reflect these name changes (e.g. m_nodes[] is now m_power_models[]).
