+++ ow_power_system/CMakeLists.txt
@@ -137,8 +137,8 @@ include_directories(
 ## With catkin_make all packages are built within a single CMake context
 ## The recommended prefix ensures that target names across packages don't collide
 add_executable(power_system_node
-  src/PowerSystemPack.cpp
   src/PowerSystemNode.cpp
+  src/PowerModelHandler.cpp
   src/PredictionHandler.cpp
   src/main.cpp)
