+++ ow_power_system/include/PowerSystemNode.h
@@ -3,10 +3,10 @@
 // this repository.
 
 // This is the header file for the PowerSystemNode class, which handles
-// the simulation of a single cell within a PowerSystemPack. It stores the
-// data used as inputs to a GSAP asynchronous prognoser and generates the next
-// input set (including any value modifications from faults or other functions) 
-// to be sent to GSAP.
+// the creation and managing of several PowerModelHandler objects (each
+// simulating part of a battery). It gathers the data used as 
+// input to GSAP's asynchronous prognosers from each PowerModelHandler, sends
+// it off, stores GSAP's predictions, and publishes them after manipulations.
 
 #ifndef __POWER_SYSTEM_NODE_H__
 #define __POWER_SYSTEM_NODE_H__
@@ -21,13 +21,30 @@
 #include <owl_msgs/BatteryStateOfCharge.h>
 #include <owl_msgs/BatteryTemperature.h>
 #include <PrognoserFactory.h>
+#include "PowerModelHandler.h"
+#include "PredictionHandler.h"
 
-// Struct that contains the information used for EoD predictions.
-struct ModelInfo {
-  double timestamp;
-  double wattage;
-  double voltage;
-  double temperature;
+#include "Messages/MessageBus.h"
+#include "Messages/ProgEventMessage.h"
+#include "Messages/ScalarMessage.h"
+#include "ModelBasedAsyncPrognoserBuilder.h"
+
+using PrognoserMap = std::map<PCOE::MessageId, PCOE::Datum<double>>;
+using PrognoserVector = std::vector<PrognoserMap>;
+
+// NOTE: This is required as a compile-time constant, so it cannot be placed in
+//       a .cfg file. The simulation must be re-built if it is changed.
+// This is the number of parallel models that will be simulated at once. The
+// expected amount for the battery pack is 24.
+const int NUM_MODELS = 24;
+
+// Struct that groups the variables/classes used to handle PowerModelHandlers.
+struct PowerModel {
+  std::string name;
+  PowerModelHandler model;
+  MessageBus bus;
+  ModelInfo model_info;
+  double previous_time;
 };
 
 class PowerSystemNode
@@ -37,77 +54,116 @@ public:
   ~PowerSystemNode() = default;
   PowerSystemNode(const PowerSystemNode&) = delete;
   PowerSystemNode& operator=(const PowerSystemNode&) = delete;
-  bool Initialize();
-  double RunOnce();
-  void GetPowerStats(ModelInfo &model);
-  void applyMechanicalPower(double mechanical_power);
-  void SetHighPowerDraw(double draw);
-  void SetCustomPowerDraw(double draw);
-  void SetCustomVoltageFault(double volts);
-  void SetCustomTemperatureFault(double tmp);
+  void initAndRun();
 private:
-  bool loadSystemConfig();
-  double generateTemperatureEstimate();
-  double generateVoltageEstimate();
-  void applyValueMods(double& power, double& voltage, double& temperature);
+  bool initModels();
+  void initTopics();
+  void injectCustomFault(bool& fault_activated,
+                         const PrognoserVector& sequence,
+                         size_t& index);
+  void injectFault(const std::string& power_fault_name,
+                   bool& fault_activated);
+  void injectFaults();
+  void jointStatesCb(const sensor_msgs::JointStateConstPtr& msg);
+  PrognoserVector loadPowerProfile(const std::string& filename,
+                                   std::string custom_file);
+  bool loadCustomFaultPowerProfile(std::string path,
+                                   std::string custom_file);
+  void publishPredictions();
+  std::string setModelName(int model_num);
+  void printTimestamp(double timestamp);
+  void printPrognoserInputs(double power,
+                            double voltage,
+                            double temperature,
+                            int index);
+  void printPrognoserOutputs(double rul,
+                             double soc,
+                             double tmp,
+                             int index);
+  void printMechanicalPower(double raw, double mean);
+  void printTopics(double rul, double soc, double tmp);
 
-  double runPrognoser(double electrical_power);
+  ros::NodeHandle m_nh;                        // Node Handle Initialization
+  ros::Publisher m_mechanical_power_raw_pub;   // Mechanical Power Raw
+  ros::Publisher m_mechanical_power_avg_pub;   // Mechanical Power Averaged
+  ros::Publisher m_state_of_charge_pub;        // State of Charge Publisher
+  ros::Publisher m_remaining_useful_life_pub;  // Remaining Useful Life Publisher
+  ros::Publisher m_battery_temperature_pub;    // Battery Temperature Publisher
+  ros::Subscriber m_joint_states_sub;          // Mechanical Power Subscriber
 
-  std::chrono::time_point<std::chrono::system_clock> m_init_time;
+  // Flags that determine if debug output is printed regarding battery status
+  // during runtime.
+  // NOTE: There's no function to get boolean variables from a config, so
+  // these bools have to be set via string comparisons. Could benefit from an
+  // update if such a function is added in the future.
+  bool m_print_debug = false;
+  bool m_timestamp_print_debug = false;
+  bool m_inputs_print_debug = false;
+  bool m_outputs_print_debug = false;
+  bool m_topics_print_debug = false;
+  bool m_mech_power_print_debug = false;
 
-  // Main system configuration: these values are overriden by values
-  // in ../config/system.cfg.
-
-  double m_min_temperature;     // minimum temp = 17.5 deg. C
-  double m_max_temperature;     // maximum temp = 21.5 deg. C
-  double m_battery_lifetime;    // Estimate of battery lifetime (seconds)
-  double m_base_voltage;        // [V] estimate
-  double m_voltage_range;       // [V]
-  double m_efficiency;          // default 90% efficiency
-  double m_baseline_wattage;    // Base power drawn by continuously-running systems.
-
-  double m_current_timestamp = 0.0;
+  // system.cfg variables:
 
   // HACK ALERT.  The prognoser produced erratic/erroneous output when
-  // given too high a power input.  This made-up value protects
-  // against this, but is a temporary hack until a circuit breaker
-  // model is added to the power system, and/or the multi-pack battery
-  // model is implemented and can handle any envisioned power draw.
-  // This initial value is overriden by the system config.
-  //
-  double m_max_gsap_input_watts = 20;
+  // given too high a power input.  The value assigned to this in system.cfg,
+  // protects against this by capping the power input, but it is a temporary
+  // hack until a circuit breaker model is added to the power system, and/or 
+  // the multi-cell battery model is implemented and can handle any envisioned
+  // power draw.
+  double m_max_gsap_input_watts;
+
+  // The maximum RUL estimation output from the Monte Carlo prediction process.
+  // If the prediction hits this value, it stops immediately and returns infinity
+  // (which is processed later into the max value instead).
+  // Lower values mean faster performance in the event a RUL prediction would
+  // exceed the horizon value, but it also means the prognoser can't return RUL
+  // values higher than this.
+  int m_max_horizon_secs;
+
+  // The number of samples each model creates during the Monte Carlo prediction
+  // process. Lower values mean faster performance, but lower accuracy (needs
+  // testing for confirmation).
+  int m_num_samples;
+
+  // GSAP's cycle time.
+  double m_gsap_rate_hz;
 
   // The expected time interval between publications. It is essentially the
   // rate at which the main loop executes. It is overridden by the value in
   // system.cfg.
-  int m_time_interval = 2;
+  int m_time_interval;
 
-  // The three main statistics relating to battery model health:
-  // Power, Voltage, & Temperature.
-  // Used by the main function to get the health of all models.
-  double m_wattage_estimate;
-  double m_voltage_estimate;
-  double m_temperature_estimate;
+  // The initial power/temperature/voltage readings used as the start values for
+  // the GSAP prognosers.
+  double m_initial_power;
+  double m_initial_temperature;
+  double m_initial_voltage;
+  double m_initial_soc;
+  
+  // End system.cfg variables.
 
-  // The stored values for modifying input to GSAP.
-  double m_added_hpd = 0.0;
-  double m_added_cpd = 0.0;
-  double m_voltage_modifier = 0.0;
-  double m_temperature_modifier = 0.0;
 
-  // End main system configuration.
+  // Flag determining if the battery has reached a fail state.
+  bool m_battery_failed = false;
 
-  // Utilize a Mersenne Twister pseudo-random generation.
-  std::mt19937 m_random_generator;
+  // Create a PowerModel struct, containing all the necessary information for a
+  // model to operate, for each model.
+  PowerModel m_power_models[NUM_MODELS];
 
-  std::uniform_real_distribution<double> m_temperature_dist;
+  // The matrix used to store EoD events.
+  EoDValues m_EoD_events[NUM_MODELS];
 
-  // Flag that indicates that the prognoser is handling current batch.
-  bool m_processing_power_batch = false;
+  // Vector w/ supporting variables that stores the moving average of the
+  // past mechanical power values.
+  const int m_moving_average_window = 25;
+  std::vector<double> m_power_values;
+  size_t m_power_values_index = 0;
 
-  bool m_trigger_processing_new_power_batch = false;
-  double m_unprocessed_mechanical_power = 0.0;
-  double m_mechanical_power_to_be_processed = 0.0;
+  bool m_high_power_draw_activated = false;
+  bool m_custom_power_fault_activated = false;
+  PrognoserVector m_custom_power_fault_sequence;
+  size_t m_custom_power_fault_sequence_index = 0;
 };
 
 #endif
